+++ chart.js/chart.esm.js
@@ -1,0 +1,5 @@
+import Chartkick from "chartkick"
+import Chart from "chart.js/auto"
+import "chartjs-adapter-date-fns"
+
+Chartkick.use(Chart)