+++ highcharts/highcharts.esm.js
@@ -1,0 +1,4 @@
+import Chartkick from "chartkick"
+import Highcharts from "highcharts"
+
+Chartkick.use(Highcharts)